Nothing Much
To build you have to click on one of the color blocks in left and you can spawn them on the Grid in the middle.. To Exit the game use the Close Button which is RED Cross on the Right side of the Screen. Also you can Reload the game with the Circle Arrows next to Exit Button.